[REFACTORING] Jmap module - Size object
